branch

+++ main.cpp
@@ -1,16 +1,16 @@
 #include <iostream>
 
-long long int fp(int x, int y, int p);
+long long int fp(long long int x, long long int y, long long int p);
 
 using namespace std;
 int main() {
-    int x,y,p;
+    long long x,y,p;
     cin>>x>>y>>p;
     cout << fp(x,y,p);
     return 0;
 }
 
-long long int fp(int x, int y, int p) {
+long long int fp(long long int x, long long int y, long long int p) {
     int res=1;
     while (y!=0 ){
         if(y%2==1){
